PET-211 refactor : 로그 모듈 구조 수정
